Added more details in Asset class.
The Asset class is now complete in details for applications that need to know Asset total supply, minimum sponsored fee, height when it was issued, etc.

Also added privateKey and publicKey in plain string format for easy reference if you want to store them in some place as back up.

Added for the meantime a CustomNode placeholder in case developers are using own full node that they want to work with. I'm currently working on some modules that will require these.