Fixed TransferTransaction to calculate accurate fee
Added CalculateFee transaction extension to ensure accuracy in the transaction fee thereby avoiding run time errors if the fee passed is not enough in cases when the Asset or the Sender Address has embedded script.

It is a good practice to always call tx.CalculateFee then compare with the balance of the sender if the amount being sent + fee is <= available balance, before signing and broadcasting transactions.