Fix Fee Calculator
I enountered an issue when using TransferTransaction.CalculateFee extension. It somehow converts the fee from decimal to long two times which results in huge amount of fee. So I moved the fee calculator outside TransactionExtensions for a more accurate and scalable solution.

I also fixed the wrong return type of CalculateFee(Transaction transaction) and changed it to long instead so we can use the tool Asset.LongToAmount when converting long value to actual decimal.

Nedd to apply this fix also on Node.MassTransfer modules.